add certification.py -add package: pillow -add function: generate_certification_image
modified command.py
-add class: TimeTrace: for timer decorator
-add function: fetch_exam: for exam
-add import: orm_study.quiz.certification
-change function: fetch_quiz
-change function: select_quiz_num: add min_num parameter
-change function: solve_quiz: decorate
-change function: start

modified reader.py
-add import: orm_study.quiz.constants.TYPE
-change function: get_chapter

modified constants.py
-add dict: TYPE

modified requirements.txt
-add pillow

modified TODO.md
